fix(website): change max content length on posts and pages by setting an environment variable (#19)

+++ mango3-core/src/enums.rs
@@ -36,7 +36,7 @@ pub enum InputError {
     CantBeBlank,
     CantBePresent,
     InvalidFormat,
-    InvalidLength(Option<usize>, Option<usize>),
+    InvalidLength(Option<u32>, Option<u32>),
     IsInvalid,
 }
